refactor(motherboard): added mod motherboard
